+++ test/java/nginx/clojure/java/WSEcho.java
@@ -13,7 +13,7 @@ public class WSEcho implements NginxJavaRingHandler {
 	@Override
 	public Object[] invoke(Map<String, Object> request) {
 		NginxJavaRequest r = (NginxJavaRequest)request;
-		NginxHttpServerChannel sc = r.handler().hijack(r, true);
+		NginxHttpServerChannel sc = r.hijack(true);
 		sc.addListener(sc, new MessageAdapter<NginxHttpServerChannel>() {
 			int total = 0;
 			@Override